fix: clear CodeQL Python quality alerts
